requirements.txt
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
DBUtils==2.0
Flask==1.1.2
Flask-MySQLdb==0.2.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
pycparser==2.20
PyMySQL==1.0.2
python-dotenv==0.15.0
Werkzeug==1.0.1
